+++ components/atoms/index.js
@@ -1,2 +1,3 @@
 export { default as View } from './View'
 export { default as Image } from './Image'
+export { default as Switch } from './Input'
