+++ components/molecules/index.js
@@ -1,1 +1,1 @@
-export { default } from './Header'
+export { default as Navbar } from './Navbar'
