+++ pages/index.js
@@ -1,4 +1,4 @@
-import Header from '../components/molecules/Header'
+import Header from '../components/organisme/Header'
 
 export default function Home() {
   return (
